Move some utils to general helpers
